[interface] Interface for master move method of game
After discussion we've decided to make a master game object that tracks all
state including which player is active and what moves are valid. The next step
would be to have a master method for making all moves that can take an
arbitrary move, validate it, and return the next game state. We need an
interface for such a move.

Closes #17